department professional roles UI

+++ SM.DTOs/Dto/FacultySimRoleDto.cs
@@ -10,7 +10,7 @@ namespace SM.Dto
         public Guid Id { get; set; }
         public string Description { get; set; }
         public Guid CourseTypeId { get; set; }
-        public byte Order { get; set; }
+        public int Order { get; set; }
 
         public virtual CourseTypeDto CourseType { get; set; }
         public virtual ICollection<CourseScenarioFacultyRoleDto> CourseScenarioFacultyRoles { get; set; }
